Internal

_PackageBundler

JuliaHub._PackageBundler.bundle — Function

bundle(dir; output = "",  force=false, allownoenv=false, verbose = true)

Creates a .tar file with the contents of dir as well as any packages that are either tracked by path (developed) outside dir or packages that are not tracked by the PkgServer. Artifacts are also bundled. The bundled packages and artifacts go into a .bundled/depot directory and is set up like a depot and can thus be made available by adding it to DEPOT_PATH.

.git and globs listed in .juliabundleignore are excluded form the bundle.

JuliaHub._PackageBundler.path_filterer — Function

path_filterer(top)

Returns a function that takes a file or directory path and checks whether that is excluded by the nearest .juliabundleignore file. The function will also ignore any .git files and directories.

The top argument specifies the highest directory up the tree that will be searched for the .juliabundleignore file.

The function will return false for any excluded files and true otherwise, and can be used as a predicate for filtering files that should be bundled.

Index

JuliaHub._PackageBundler.bundle

JuliaHub._PackageBundler.path_filterer