Exceptions

JuliaHub.jl is designed in a way that the only errors it should throw under normal circumstances are subtypes of JuliaHubException (in addition to standard ArgumentErrors and MethodErrors etc. from invalid function calls). Any unhandled errors from JuliaHub.jl or its dependencies should be considered a bug.

Debugging JuliaHub.jl issues

You can also enable debug logging for JuliaHub, which will make JuliaHub.jl print out additional debug messages, by setting the JULIA_DEBUG environment variable:

ENV["JULIA_DEBUG"]="JuliaHub"

JuliaHub.JuliaHubException — Type

abstract type JuliaHubException <: Exception

Abstract supertype of all JuliaHub.jl exception types.

JuliaHub.AuthenticationError — Type

struct AuthenticationError <: JuliaHubException

Exception thrown if the authentication fails. The .msg fields contains a human-readable error message.

JuliaHub.InvalidAuthentication — Type

struct InvalidAuthentication <: JuliaHubException

This exception is thrown if the authentication token is invalid or has expired. Re-authenticating with JuliaHub.authenticate should generally be sufficient to resolve the issue.

JuliaHub.InvalidRequestError — Type

struct InvalidRequestError <: JuliaHubException

An exception thrown if the request was rejected by the backend due to request parameters that are inconsistent with the backend state. The .msg field contains the error message.

JuliaHub.JuliaHubConnectionError — Type

struct JuliaHubConnectionError <: JuliaHubException

An exception thrown if there is a communication error with JuliaHub.

The .msg field contains the error message. If there is an underlying exception, it is stored in the .exception field.

JuliaHub.JuliaHubError — Type

struct JuliaHubError <: JuliaHubException

An exception thrown if there is an unexpected response from or backend failure in JuliaHub.

The .msg field contains the error message. If there is an underlying exception, it is stored in the .exception field.

JuliaHub.PermissionError — Type

struct PermissionError <: JuliaHubException

Thrown if the currently authenticated user does not have the necessary permissions to perform the operation. The .msg field contains the error message, and .response may contain the raw server response.

JuliaHub.ProjectNotSetError — Type

struct ProjectNotSetError <: JuliaHubException

Exception thrown by a project-related operation that requires a project to be specified, but neither an explicit project reference was provided, nor was the project set for the authentication object.

JuliaHub.InvalidJuliaHubVersion — Type

struct InvalidJuliaHubVersion <: JuliaHubException

Thrown if the requested operation is not supported by the JuliaHub instance. .msg contains a more detailed error message.

Tip

This generally means that the functionality you are attempting to use requires a newer JuliaHub version.

Index

JuliaHub.AuthenticationError

JuliaHub.InvalidAuthentication

JuliaHub.InvalidJuliaHubVersion

JuliaHub.InvalidRequestError

JuliaHub.JuliaHubConnectionError

JuliaHub.JuliaHubError

JuliaHub.JuliaHubException

JuliaHub.PermissionError

JuliaHub.ProjectNotSetError