Experimental APIs

🐉 Hic Sunt Dracones

The JuliaHub.Experimental module contains various experimental APIs.

JuliaHub.Experimental — Module

module Experimental

Home for experimental JuliaHub.jl APIs.

Unstable APIs

These APIs are considered highly unstable. Both JuliaHub platform version changes, and also JuliaHub.jl package changes may break these APIs at any time. Depend on them at your own peril.

Reference

JuliaHub.Experimental.Registry — Type

struct Registry

Represents a Julia package registry on JuliaHub.

Unstable API

This API is not part of the public API and does not adhere to semantic versioning.

This APIs is considered highly unstable. Both JuliaHub platform version changes, and also JuliaHub.jl package changes may break it at any time. Depend on it at your own peril.

JuliaHub.Experimental.registries — Function

JuliaHub.Experimental.registries() -> Vector{Experimental.Registry}

Return the list of registries configured on the JuliaHub instance.

Unstable API

This API is not part of the public API and does not adhere to semantic versioning.

This APIs is considered highly unstable. Both JuliaHub platform version changes, and also JuliaHub.jl package changes may break it at any time. Depend on it at your own peril.

JuliaHub.Experimental.register_package — Function

JuliaHub.Experimental.register_package(
    auth::Authentication,
    registry::Union{AbstractString, Registry},
    repository_url::AbstractString;
    # Optional keyword arguments:
    [notes::AbstractString,]
    [branch::AbstractString,]
    [subdirectory::AbstractString,]
    [git_server_type::AbstractString]
) -> String | Nothing

Initiates a registration PR of the package at repository_url in Returns the URL of the registry PR, or nothing if the registration failed.

Example

using JuliaHub
auth = JuliaHub.authenticate("juliahub.com")
JuliaHub._registries(auth)

r = JuliaHub.Experimental.register_package(
    auth,
    "MyInternalRegistry",
    "https://github.com/MyUser/MyPackage.jl";
    notes = "This was initiated via JuliaHub.jl",
)

Unstable API

This API is not part of the public API and does not adhere to semantic versioning.

This APIs is considered highly unstable. Both JuliaHub platform version changes, and also JuliaHub.jl package changes may break it at any time. Depend on it at your own peril.