Projects

These APIs allow you to interact with JuliaHub projects: the datasets that have been attached to them, and project deployments.

project_datasets and project_dataset let you list and access datasets linked to a project

upload_project_dataset allows uploading new versions of project-linked datasets

project_deployment_specs and deploy_project let you list a deployable project's deployment specifications and start new deployments

Automatic project authentication

The Authentication object can be associated with a default project UUID, which will then be used to for all project operations, unless an explicit project gets passed to override the default.

Importantly, JuliaHub.authenticate will automatically pick up the the JuliaHub project UUID from the JULIAHUB_PROJECT_UUID environment variable. This means in JuliaHub cloud jobs and IDEs, it is not necessary to manually set the project, and JuliaHub.jl will automatically. However, you can opt-out of this behavior by explicitly passing a project=nothing to JuliaHub.authenticate.

You can always verify that your operations are running in the context of the correct project by checking the Authentication object, e.g. via current_authentication:

julia> JuliaHub.current_authentication()
JuliaHub.Authentication("https://juliahub.com", "username", *****; project_id = "cd6c9ee3-d15f-414f-a762-7e1d3faed835")

Reference

JuliaHub.project_datasets — Function

JuliaHub.project_datasets([project::ProjectReference]; [auth::Authentication]) -> Vector{Dataset}

Returns the list of datasets attached to the project, as a list of Dataset objects. If the project is not explicitly specified, it uses the project of the authentication object.

May throw a ProjectNotSetError. Will throw an [InvalidRequestError] if the currently authenticated user does not have access to the project or the project does not exists.

julia> JuliaHub.current_authentication()
JuliaHub.Authentication("https://juliahub.com", "username", *****; project_id = "cd6c9ee3-d15f-414f-a762-7e1d3faed835")

julia> JuliaHub.project_datasets()
3-element Vector{JuliaHub.Dataset}:
 JuliaHub.project_dataset(("username", "example-dataset"); project="cd6c9ee3-d15f-414f-a762-7e1d3faed835")
 JuliaHub.project_dataset(("anotheruser", "publicdataset"); project="cd6c9ee3-d15f-414f-a762-7e1d3faed835")
 JuliaHub.project_dataset(("username", "blobtree/example"); project="cd6c9ee3-d15f-414f-a762-7e1d3faed835")

JuliaHub.project_dataset — Function

JuliaHub.project_dataset(dataset::DatasetReference; [project::ProjectReference], [auth]) -> Dataset

Looks up the specified dataset among the datasets attached to the project, returning a Dataset object, or throwing an InvalidRequestError if the project does not have such dataset attached.

julia> JuliaHub.project_dataset(("username", "blobtree/example"))
Dataset: blobtree/example (BlobTree)
 owner: username
 description: An example dataset
 versions: 1
 size: 57 bytes
 tags: tag1, tag2
 project: cd6c9ee3-d15f-414f-a762-7e1d3faed835 (not writable)

Implicit dataset owner

When passing just the dataset name for dataset (i.e. <: AbstractString), then, just like for the non-project JuliaHub.dataset function, it is assumed that the owner of the dataset should be the currently authenticated user.

However, a project may have multiple datasets with the same name attached to it (if they are owned by different users). The best practice when accessing datasets in the context of projects is to fully specify their name (i.e. also include the username).

Non-dynamic dataset objects

Dataset objects represents the dataset metadata when the Julia object was created (e.g. with dataset), and are not automatically kept up to date. To refresh the dataset metadata, you can pass an existing Dataset object to JuliaHub.dataset or project_dataset.

JuliaHub.upload_project_dataset — Function

JuliaHub.upload_project_dataset(
    dataset::DatasetReference, local_path;
    progress=true,
    [project::ProjectReference],
    [auth::Authentication],
) -> Dataset

Uploads a new version of a project-linked dataset.

By default, the new dataset version will be associated with the project of the current authentication session (if any), but this can be overridden by passing project.

Permissions

Note that in order for this to work, you need to have edit rights on the projects and the dataset needs to have been marked writable by the dataset owner. However, unlike for normal datasets uploads (with upload_dataset), you do not need to be the dataset owner to upload new versions.

Tip

The function call is functionally equivalent to the following upload_dataset call

JuliaHub.upload_dataset(
    dataset, local_path;
    create=false, update=true, replace=false,
)

except that the upload is associated with a project.

JuliaHub.ProjectReference — Type

const ProjectReference :: Type

Type constraint on the argument that specifies the project in projects-related APIs that (e.g. project_datasets).

Presently, you can specify the project by directly passing the project UUID. The UUID should be either a string (<: AbstractString) or an UUIDs.UUID object.

Project deployments

JuliaHub.project_deployment_specs — Function

JuliaHub.project_deployment_specs([project::ProjectReference]; [auth::Authentication]) -> Vector{ProjectDeploymentSpec}

Returns the list of deployment specifications (ProjectDeploymentSpec objects) of the project. If project is omitted, the project associated with the authentication object is used.

Throws an InvalidRequestError if the project does not exist, if the user does not have owner or editor access to it, or if the project is not a deployable project.

julia> JuliaHub.project_deployment_specs()
2-element Vector{JuliaHub.ProjectDeploymentSpec}:
 JuliaHub.ProjectDeploymentSpec(33, "Default")
 JuliaHub.ProjectDeploymentSpec(96, "Large")

JuliaHub.deploy_project — Function

JuliaHub.deploy_project(
    [project::ProjectReference];
    [spec::Union{ProjectDeploymentSpec, Integer, AbstractString}],
    [auth::Authentication]
) -> Job

Starts a new deployment of the project, according to one of its deployment specifications, and returns the corresponding Job object. If project is omitted, the project associated with the authentication object is used.

The deployment specification can be specified with spec, either as a ProjectDeploymentSpec object, its numeric ID, or its name (see project_deployment_specs). If spec is omitted, the project's default deployment specification is used. If the project has no default specification, but has exactly one specification, that one is used. Otherwise an InvalidRequestError is thrown.

Also throws an InvalidRequestError if the project does not exist, if the user does not have owner or editor access to it, if the project is not a deployable project, or if the specified deployment specification does not exist.

Note

JuliaHub bundles the project before queueing the deployment job, so it can take a little while for the job to become visible. This function blocks until the job can be queried (up to a few minutes), and throws a JuliaHubError with the job name if the job does not show up in that time.

julia> job = JuliaHub.deploy_project(; spec="Large")
JuliaHub.Job: jr-xf4tslavut (Completed)
 submitted: 2023-03-15T07:56:50.974+00:00
 started:   2023-03-15T07:56:51.251+00:00
 finished:  2023-03-15T07:56:59.000+00:00
 files:
  - code.jl (input; 3 bytes)
  - code.jl (source; 3 bytes)
  - Project.toml (project; 244 bytes)
  - Manifest.toml (project; 9056 bytes)
 outputs: "{}"

Non-dynamic job objects

Job objects represents the jobs when the objects were created (e.g. by a job function) and are not automatically kept up to date. To refresh the job information, you can pass the existing Job to JuliaHub.job before passing it to this function.

JuliaHub.ProjectDeploymentSpec — Type

struct ProjectDeploymentSpec

Represents a deployment specification of a deployable JuliaHub project. Deployment specifications are managed in the JuliaHub web UI, and can be listed with project_deployment_specs. A project can be deployed according to a specification with deploy_project.

Objects have the following properties:

id :: Int: the numeric ID of the specification (unique across the JuliaHub instance)

project_id :: UUID: the UUID of the project the specification belongs to

name :: String: the (project-unique) name of the specification

description :: String: description of the specification

machine_type :: String: the JuliaHub machine type the deployment runs on (e.g. "m64")

port :: Int: the port the deployed application listens on

authentication :: String: how access to the deployment is authenticated (e.g. "me", "password")

sysimage_build :: Bool: whether a system image is built for the deployment

default :: Bool: whether this is the default deployment specification of the project

Index